made some progress, but overall does not work
