+++ src/lib.rs
@@ -1,9 +1,2 @@
 pub mod linear_model;
 pub mod tensor;
-// fn main() {
-// let v = vec![0, 1, 2, 3];
-// let matrix = Tensor::new(v, vec![2, 2]);
-
-// println!("Hello, world!");
-// println!("Matrix: \n{:?}", matrix);
-// }
